Changes to all setup files to make robust to repeat usage, some bugfixes

+++ scripts/steps/04-open-drone-server.sh
@@ -4,8 +4,13 @@ source "$(dirname "${BASH_SOURCE[0]}")/../env.sh"
 
 # Clone rest API
 cd $HOME/src/
-git clone https://github.com/robotics-88/open-drone-server.git
+if [ -d "open-drone-server" ]; then
+    echo "Directory open-drone-server already exists. Skipping clone."
+else
+    git clone https://github.com/robotics-88/open-drone-server.git
+fi
 cd open-drone-server
+git pull
 python3 -m venv .env
 source .env/bin/activate
 pip install -r requirements.txt
@@ -13,6 +18,10 @@ deactivate
 
 
 # Create systemd service for rest backend
+if [ -f "/etc/systemd/system/rest-server-drone.service" ]; then
+    echo "Systemd service file already exists. Skipping creation. ✅ 04 Success"
+    exit 0
+fi
 echo "Creating systemd service for open-drone-server..."
 cat <<EOF | sudo tee /etc/systemd/system/rest-server-drone.service > /dev/null
 [Unit]
@@ -34,3 +43,5 @@ EOF
 sudo systemctl daemon-reload
 sudo systemctl enable rest-server-drone.service
 sudo systemctl start rest-server-drone.service
+
+echo "04 Open Drone Server setup completed. ✅ Success"
